+++ ChangeOwner/AbstractHandler.php
@@ -78,10 +78,8 @@ abstract class AbstractHandler
      */
     public function getOldOwner(Entity $content): UserEntity
     {
-        $oldContentOwner = null;
         try
         {
-            /** @noinspection PhpUndefinedFieldInspection */
             $oldContentOwner = $content->User;
         }
         catch (\InvalidArgumentException $e)
@@ -229,7 +227,6 @@ abstract class AbstractHandler
 
         return \XF::asVisitor($newOwner, function () use($content, $error)
         {
-            /** @noinspection PhpUndefinedMethodInspection */
             return $content->canView($error);
         });
     }