+++ Entity/ContentTrait.php
@@ -1,5 +1,7 @@
 <?php
 
+/** @noinspection PhpMultipleClassDeclarationsInspection */
+
 namespace TickTackk\ChangeContentOwner\Entity;
 
 use TickTackk\ChangeContentOwner\ChangeOwner\AbstractHandler;
@@ -15,9 +17,10 @@ trait ContentTrait
      * @param bool $throw
      *
      * @return AbstractHandler
+     *
      * @throws \Exception
      */
-    public function getChangeOwnerHandler(bool $throw = false) : AbstractHandler
+    public function getChangeOwnerHandler(bool $throw = false) :? AbstractHandler
     {
         $contentType = $this->getEntityContentType();
         $handlerClass = \XF::app()->getContentTypeFieldValue($contentType, 'change_owner_handler_class');