+++ Service/Content/AbstractOwnerChanger.php
@@ -1,5 +1,7 @@
 <?php
 
+/** @noinspection PhpMultipleClassDeclarationsInspection */
+
 namespace TickTackk\ChangeContentOwner\Service\Content;
 
 use TickTackk\ChangeContentOwner\ChangeOwner\AbstractHandler;
@@ -269,7 +271,7 @@ abstract class AbstractOwnerChanger extends AbstractService
             throw new \InvalidArgumentException('Invalid hour provided.');
         }
 
-        if ($newTime['minute'] < 0 || $newTime['hour'] > 59)
+        if ($newTime['minute'] < 0 || $newTime['minute'] > 59)
         {
             throw new \InvalidArgumentException('Invalid minute provided.');
         }
@@ -382,23 +384,21 @@ abstract class AbstractOwnerChanger extends AbstractService
         {
             if ($this->contentNewDateCounter === null) // first content
             {
-                $this->contentNewDateCounter = 0;
+                $this->contentNewDateCounter = 1;
             }
-            else
-            {
-                foreach ($timeIntervals AS $unit => $value)
-                {
-                    if (!$value)
-                    {
-                        continue;
-                    }
 
-                    $counter = $value + $this->contentNewDateCounter;
-                    $dateTime->modify("+{$counter} {$unit}");
+            foreach ($timeIntervals AS $unit => $value)
+            {
+                if (!$value)
+                {
+                    continue;
                 }
 
-                $this->contentNewDateCounter++;
+                $counter = $value * $this->contentNewDateCounter;
+                $dateTime->modify("+{$counter} {$unit}");
             }
+
+            $this->contentNewDateCounter++;
         }
 
         $this->contentNewDateMapping[$uniqueKey] = $dateTime->getTimestamp();
@@ -654,7 +654,7 @@ abstract class AbstractOwnerChanger extends AbstractService
             {
                 $handler = $this->getHandler();
 
-                foreach ($this->contents AS $id => $content)
+                foreach ($this->contents AS $content)
                 {
                     if (!$handler->canNewOwnerViewContent($content, $newOwner, $error))
                     {
@@ -745,7 +745,7 @@ abstract class AbstractOwnerChanger extends AbstractService
         $logger = $this->app->logger();
         $logModerator = $this->getLogModerator();
 
-        foreach ($this->contents AS $id => $content)
+        foreach ($this->contents AS $content)
         {
             $this->additionalEntitySave($content);
 
@@ -797,7 +797,6 @@ abstract class AbstractOwnerChanger extends AbstractService
     /**
      * @param Entity|ContentEntityInterface $content
      *
-     * @throws \XF\Db\Exception
      * @throws \XF\PrintableException
      * @throws \Exception
      */