+++ _output/entity_hint.php
@@ -5,7 +5,6 @@
 
 namespace XF\Entity
 {
-    /** @noinspection PhpUndefinedClassInspection */
     /** @noinspection PhpMissingDocCommentInspection */
     class LikedContent extends \XF\Mvc\Entity\Entity {}
 }